Refactor util.hpp and BhArray.hpp
