+++ bridge/cxx/include/bhxx/util.hpp
@@ -1,6 +1,3 @@
-//
-// Copyright (C) 2017 by the linalgwrap authors
-//
 /*
 This file is part of Bohrium and copyright (c) 2012 the Bohrium
 team <http://www.bh107.org>.
@@ -22,49 +19,13 @@ If not, see <http://www.gnu.org/licenses/>.
 */
 
 #pragma once
-
-#include <bh_view.hpp>
-#include <functional>
-#include <numeric>
-#include <ostream>
-#include <vector>
+#include "BhArray.hpp"
 
 namespace bhxx {
 
-template <typename T, std::size_t MaxLength>
-struct SVector : public std::vector<T> {
-  public:
-    using std::vector<T>::vector;
-    SVector(const std::vector<T>& other) : SVector(other.begin(), other.end()) {}
-    SVector() = default;
-
-    T sum() const { return std::accumulate(this->begin(), this->end(), T{0}); }
-    T prod() const {
-        return std::accumulate(this->begin(), this->end(), T{1}, std::multiplies<T>());
-    }
-};
-
-// Some common SVectors
-typedef SVector<int64_t, BH_MAXDIM> Stride;
-typedef SVector<size_t, BH_MAXDIM>  Shape;
-
-// Return a contiguous stride (row-major) based on `shape`
-Stride contiguous_stride(const Shape& shape);
-
-template <typename T, std::size_t MaxLength>
-std::ostream& operator<<(std::ostream& o, const SVector<T, MaxLength>& vec) {
-    o << '(';
-
-    if (!vec.empty()) {
-        auto it = std::begin(vec);
-        o << *it;
-        ++it;
-
-        for (; it != std::end(vec); ++it) o << ',' << *it;
-    }
-
-    o << ')';
-    return o;
-}
+/** Convert an array with exactly one element to a scalar by calling
+ *  sync and flush and returning the value */
+template <typename T>
+T as_scalar(BhArray<T>& ary);
 
 }  // namespace bhxx
