+++ bridge/cxx/src/util.cpp
@@ -18,21 +18,56 @@ GNU Lesser General Public License along with Bohrium.
 If not, see <http://www.gnu.org/licenses/>.
 */
 
+#include <bhxx/Runtime.hpp>
+#include <bhxx/array_operations.hpp>
 #include <bhxx/util.hpp>
-
-using namespace std;
+#include <complex>
 
 namespace bhxx {
 
-// Return a contiguous stride (row-major) based on `shape`
-Stride contiguous_stride(const Shape& shape) {
-    Stride  ret(shape.size());
-    int64_t stride = 1;
-    for (int64_t i = shape.size() - 1; i >= 0; --i) {
-        ret[i] = stride;
-        stride *= static_cast<int64_t>(shape[i]);
+template <typename T>
+T as_scalar(BhArray<T>& ary) {
+    if (ary.base == nullptr) {
+        throw std::runtime_error(
+              "Cannot call bhxx::as_scalar on BhArray objects without base");
     }
-    return ret;
+
+    if (ary.n_elem() != 1) {
+        throw std::runtime_error(
+              "Cannot call bhxx::as_scalar on BhArray objects with more than one "
+              "element");
+    }
+
+    sync(ary);
+    Runtime::instance().flush();
+
+    const T* data = ary.data();
+    if (data == nullptr) {
+        throw std::runtime_error("Cannot get the scalar from an uninitialised BhArray.");
+    }
+
+    return *data;
 }
+
+
+// Instantiate all possible types of `BhArray`
+#define INSTANTIATE(T)                          \
+    template T          as_scalar(BhArray<T>&); \
+
+INSTANTIATE(bool);
+INSTANTIATE(int8_t);
+INSTANTIATE(int16_t);
+INSTANTIATE(int32_t);
+INSTANTIATE(int64_t);
+INSTANTIATE(uint8_t);
+INSTANTIATE(uint16_t);
+INSTANTIATE(uint32_t);
+INSTANTIATE(uint64_t);
+INSTANTIATE(float);
+INSTANTIATE(double);
+INSTANTIATE(std::complex<float>);
+INSTANTIATE(std::complex<double>);
+
+#undef INSTANTIATE
 
 }  // namespace bhxx
